test(seed): add integrity checks for templates and required metadata
